remove OK dialog on artwork write error
on batch runs, this makes the game import at every such error. the operator then needs to wake up, hit enter, and wait for the next error to come up, which is really annoying. the dialog is useless anyways because it doesn't say what the error is, it just says to look in the log... which we can do anyways, without the error message.